[ALPS02353522] Display: fix the coverity issue-0
[Detail] 1. fbconfig_kdebug.c
	 2. primary_display.c

MTK-Commit-Id: 7b5c3a98b10d948e10747efef09d7fe918218443

CR-Id: ALPS02353522
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6763/videox/primary_display.c
@@ -5831,7 +5831,7 @@ static int smart_ovl_try_switch_mode_nolock(void)
 	data_config = dpmgr_path_get_last_config(disp_handle);
 
 	/* calc wdma/rdma data size */
-	rdma_sz = data_config->dst_h * data_config->dst_w * 3;
+	rdma_sz = (unsigned long long)data_config->dst_h * data_config->dst_w * 3;
 
 	/* calc ovl data size */
 	ovl_sz = 0;
@@ -5841,7 +5841,7 @@ static int smart_ovl_try_switch_mode_nolock(void)
 		if (ovl_cfg->layer_en) {
 			unsigned int Bpp = UFMT_GET_Bpp(ovl_cfg->fmt);
 
-			ovl_sz += ovl_cfg->dst_w * ovl_cfg->dst_h * Bpp;
+			ovl_sz += (unsigned long long)ovl_cfg->dst_w * ovl_cfg->dst_h * Bpp;
 		}
 	}
